Allow non-configurable demos
This allows demos to be non-configurable. Demos with no `variants`
property in the demo defaults will now be generated correctly, but they
will not include the sidebar. We may want to have non-customisable demos
in the future, and this also aids development as you can get something
up-and-running sooner.

I've moved the buttons in `demo-area.js` to avoid complex nested
ternaries. I have no idea whether this is Reacty, but named variables
makes things easier to keep track of 🤷‍♂️